pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reg-policy-weekly-report-skill-lite"
version = "0.1.0"
description = "平安集团《监管政策周报》生产 Skill（轻量版）：政策清单 JSON → 终稿口径 Word，附 few-shot / forbidden expansion 自进化闭环。"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
    { name = "Compliance & Coordination, Ping An Group (Internal)" },
]
keywords = [
    "regulatory",
    "policy",
    "weekly-report",
    "compliance",
    "ping-an",
    "llm",
    "skill",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.6,<3.0",
    "pyyaml>=6.0,<7.0",
    "httpx>=0.27,<1.0",
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=5.1,<6.0",
    "python-docx>=1.1,<2.0",
    "anthropic>=0.40,<1.0",
    "click>=8.1,<9.0",
    "rich>=13.7,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-mock>=3.12,<4.0",
]

[project.scripts]
reg-policy-weekly = "src.main:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.md", "*.docx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
]
